Move execution introspection methods to separate concern
